Add api to extend the idle timeout for jobs.
We have had issues with builders stopping after 10 minutes, while
builds are still running. Builds do not consistently hit the proxy
endpoing that extends the idle timeout, and we are not monitoring the
traffic on the tcp streams used to marshall the docker build.

Add a new endpoint /flyio/v1/extendDeadline, which will reset the
docker job idle timeout to the full 10 minutes each time it is
called. flyctl will call this endpoint periodically while the image is
building. This will ensure the builder machine does not get stopped
before the build finishes.